Provides a flag to execute UnixBench with the > 16 cores patch
UnixBench by default is designed to only execute up to 16 jobs
concurrently. If your system has more than 16 CPUs available, UnixBench
will not execute enough jobs to fully utilize all the CPUs in the machine,
which will result in a lower score that what the system is capable of.
A patch was submitted to the UnixBench issues that addresses this issue.
This commit allows the user to optionally use this patch by setting the
following flag:

  --unixbench_all_cores

Official link to issue:
https://code.google.com/p/byte-unixbench/issues/detail?id=4