Kubernetes provider.
Instead of VMs in Kubernetes we have a concept of PODs. Create() method
creates a POD with a single container inside. Access to this container is
done through the Service. Authentication is done via Secret - SSH key is
base64 encoded and stored in a file which is later mounted as a volume.

Additionaly, Ceph support was added. You can specify IP addresses of Ceph
Monitors. When scratch disk is needed, RBD image will be created with ext4
filesystem. This volume has to be created _before_ the actual POD launch.
Ceph volume is currently the only option if the benchmark requires scratch
disk but I plan to add local disk support in the next check-ins.

Signed-off-by: Mateusz Blaszkowski <mateusz.blaszkowski@intel.com>