Handle User Job Files Better in fio_benchmark.py
Given a user job file, we either remove their filename parameter and
provide our own (if testing against a raw device) or keep user
filenames but use --directory to make sure the files are on the device
we are testing (if using a filesystem). Also remove the templating
functionality that the previous commit added to job files. This is a
different fix for the same bug.