Create BenchmarkConfigSpec to aggregate benchmark input checking.
Part of a solution to #787.
Creating a BenchmarkConfigSpec performs verification on inputs for a
benchmark run, including creation of sub-spec objects like the VM and
disk specs and the input-checking that they contain. In a separate PR,
the creation of the BenchmarkConfigSpec will be moved out of the
BenchmarkSpec.__init__ to before the call to RunBenchmark so that the
inputs of all configured benchmark runs can be verified before any
benchmark is started.