Factor object_storage_api_tests.py with separate backend objects
Move the code for each object storage provider in
object_storage_api_tests.py to a separate per-provider "backend" object,
and update the benchmarks to work with this system. Add a new
TestBackend scenario for testing these backends (clearly marked as for
debugging only in the docstring). Also take the opportunity to improve
testing of the benchmarks, by implementing a mock backend and adding
some integration tests that use it.