Fixes race condition in Aerospike benchmark.
The fucntion ConfigureAndStart(..) installs and starts the Aerospike server
onto the target VM. Once this finishes the framework moves on and tries to
start the benchmarking, and enters a race between the clients attempting to
connect to the server and the server finishing its initialization.

Previously this was solved by a sleep(5) at the end of this function, however
this was insufficient for most of the runs I attempted. Increasing this to
sleep(15) did work for me, but an actual check seemed superior.

This change adds a proactive check that the server is up and ready for
connections before returning from the "ConfigureAndStart" function. We
accomplish this by sshing into the server vm, and connecting to  Aerospike's
out of band telnet management interface on the internal IP address and issueing
 the `status` command. Using the @Retry decorator from vm_util.py we
continuously check until a 5 minute timeout.

See the aerospike documentation for more info on the status check.
  http://www.aerospike.com/docs/reference/info#status
  http://www.aerospike.com/docs/tools/asinfo#aerospike-s-telnet-port