Add support for GPUs in GKE.
Change cuda_toolkit_installtion_dir to be a flag, and derive the path of nvidia-smi from this path (removes requirement that nvidia-smi is in PATH).

Modify kubernetes_virtual_machine so that it installs 'sudo' on the nvidia/cuda image.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=175030157