Initial support for downloading pre-provisioned data into a VM.
This change adds a mechanism to pull data that has been pre-provisioned in a particular cloud into VMs running in that cloud, and adds implementations for GCP and AWS. The mechanism validates the downloaded data based on an md5sum that is stored alongside the data file in the benchmark code.

The provider VM implementations must override DownloadPreprovisionedBenchmarkData. For a particular OS, the mixin must override CheckPreprovisionedBenchmarkData to validate the md5sum.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=181805107