Add flag for number of MPI process(es) to use in the OMB benchmark.
Some benchmarks only (example: "bw") only support 1 thread per host, skip requests for that benchmark with more than that.

Order in the default list is [1, 0] so that the run with 1 thread per host happens first followed by a run with the number of threads per host equal to the number of cores per host.

PiperOrigin-RevId: 360250895