Only install pip and pip3 when it they not distributed with Python. When it needs to be installed, install the appropriate version from get-pip.py.
Previously we installed pip/pip3 using potentially old Linux packages and then trying to upgrade to HEAD. This broke when the old pip either failed to talk to PyPI: https://status.python.org/incidents/grk0k7sz6zkp
or when it didn't know a safe version of pip to upgrade to.
get-pip.py avoids this by installing the correct version for this Python version.

PiperOrigin-RevId: 378512225