GCE: Change WaitUntilRunning to poll on the `instances create` operation rather than the VM instance itself.
Error parsing in the create phase is currently reliant on the information is returned by a synchronous create command. Now that we use the `--async` flag, this error handling is invalid because the create command returns before the process completes. This leads to provisioning failures whose causes are difficult to pin down.

To ensure errors are properly classified, I refactored the error handling from Create into a separate function that can be invoked in the create phase and in the WaitUntilRunning phase.

PiperOrigin-RevId: 524400385